fix(rust): `oneOf` generation for client
Solves #17869 and #17896 and also includes unmerged $17898

Unfortunately it affects quite a lot of code, but we can see that only client-side models were affected by re-generation.
I tried to split this PR to several, but they're really coupled and hard to create a chain of PRs.

+++ samples/client/others/rust/reqwest/oneOf/src/models/pizza.rs
@@ -1,3 +1,5 @@
+use crate::models;
+
 /*
  * ByRefOrValue
  *
@@ -14,7 +16,7 @@
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Pizza {
     #[serde(rename = "pizzaSize", skip_serializing_if = "Option::is_none")]
-    pub pizza_size: Option<f32>,
+    pub pizza_size: Option<f64>,
     /// Hyperlink reference
     #[serde(rename = "href", skip_serializing_if = "Option::is_none")]
     pub href: Option<String>,
